Ex (3.5) 7(fix for(size_t to int))

+++ С++/The C++ programming language/Ex (3.5) 7.cpp
@@ -1,5 +1,6 @@
 #include<iostream>
 #include<string.h>
+
 using namespace std;
 
 size_t mstrlen(const char* s)
@@ -25,7 +26,7 @@ int mstrcmp(const char* s1, const char* s2)
 {
 	if(strlen(s1) == strlen(s2))
 	{
-		for(int i = 0; i <= strlen(s1);i++)
+		for(size_t i = 0; i <= strlen(s1);i++)
 		{
 			if(*s1 != *s2) return 0;
 			s1++; s2++;
